Voter class added.

+++ src/Controller/ContactController.php
@@ -6,17 +6,26 @@ namespace App\Controller;
 
 use App\Entity\Contact;
 use App\Repository\ContactRepository;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 class ContactController extends AbstractController
 {
-
+    /**
+     * @IsGranted("ROLE_USER")
+     * @param \App\Repository\ContactRepository $contactRepository
+     *
+     * @return \Symfony\Component\HttpFoundation\Response
+     *
+     */
     public function index(ContactRepository $contactRepository)
     {
         return $this->render('contact/index.html.twig', [
-            'contacts' => $contactRepository->findAll(),
+            'contacts' => $contactRepository->findBy([
+                'user' => $this->getUser(),
+            ]),
         ]);
     }
 
@@ -27,12 +36,15 @@ class ContactController extends AbstractController
      */
     public function show(Contact $contact)
     {
+        $this->denyAccessUnlessGranted('VIEW', $contact);
+
         return $this->render('contact/show.html.twig', [
             'contact' => $contact,
         ]);
     }
 
     /**
+     * @IsGranted("ROLE_USER")
      * @return \Symfony\Component\HttpFoundation\Response
      */
     public function create()
@@ -50,6 +62,8 @@ class ContactController extends AbstractController
      */
     public function edit(Contact $contact)
     {
+        $this->denyAccessUnlessGranted('UPDATE', $contact);
+
         return $this->render('contact/edit.html.twig', [
             'contact' => $contact,
             'errors' => [],
@@ -57,6 +71,7 @@ class ContactController extends AbstractController
     }
 
     /**
+     * @IsGranted("ROLE_USER")
      * @param \Symfony\Component\HttpFoundation\Request                 $request
      *
      * @param \Symfony\Component\Validator\Validator\ValidatorInterface $validator
@@ -65,7 +80,6 @@ class ContactController extends AbstractController
      */
     public function store(Request $request, ValidatorInterface $validator)
     {
-
         $contact = new Contact();
         $contact->setName($request->get('name'));
         $contact->setEmail($request->get('email'));
@@ -78,7 +92,7 @@ class ContactController extends AbstractController
                 'contact' => $contact,
             ]);
         }
-
+        $contact->setUser($this->getUser());
         $entityManager = $this->getDoctrine()->getManager();
         $entityManager->persist($contact);
         $entityManager->flush();
@@ -99,6 +113,8 @@ class ContactController extends AbstractController
      */
     public function update(Contact $contact, Request $request, ValidatorInterface $validator)
     {
+        $this->denyAccessUnlessGranted('UPDATE', $contact);
+
         $contact->setName($request->get('name'));
         $contact->setEmail($request->get('email'));
         $contact->setPhoneNumber($request->get('phone_number'));
@@ -127,6 +143,8 @@ class ContactController extends AbstractController
      */
     public function destroy(Contact $contact)
     {
+        $this->denyAccessUnlessGranted('DELETE', $contact);
+
         $entityManager = $this->getDoctrine()->getManager();
         $entityManager->remove($contact);
         $entityManager->flush();
